+++ CHANGELOG.md
@@ -9,10 +9,11 @@
       - *default value* mechanism
       - *equals* operation
       - *deepcopy* support
-- **New**: Implemented abstract `Clocked` class to unite handling of registers and memories
+- **New**: Implemented `Clock` class to unite handling of registers and memories
   - Write operations on memories and register files will _not_ commit until the next clock tick
     anymore
     - This is to ensure data coherency throughout the current simulation cycle
+  - Added `Clocked` class as abstract base case for clocked elements (registers, memories)
 - **New**: Added basic logging
   - For logging purposes, the design will be scanned for submodules, ports and registers
     - The instance names of these components is then added as an attribute to each component